test: move raytracing tests to their own file
